extract LLM params parsing
